- Add field check on "down" method in "move_avatar" migration - Add reverse renaming on "down" method in "change_all_table_prefix" migration

+++ src/database/migrations/2019_07_11_231939_move_avatar_column_to_users_table.php
@@ -8,7 +8,7 @@ use App\Base\Traits\MigrateDataTenant;
 class MoveAvatarColumnToUsersTable extends Migration
 {
     use MigrateDataTenant;
-    
+
     /**
      * Run the migrations.
      *
@@ -16,27 +16,27 @@ class MoveAvatarColumnToUsersTable extends Migration
      */
     public function up()
     {
-        if(config('AppConfig.system.multitenant.active',false) && $this->tenantMigrateMode()==false){
-            if (!Schema::hasColumn('users','avatar')) {
+        if (config('AppConfig.system.multitenant.active', false) && $this->tenantMigrateMode() == false) {
+            if (!Schema::hasColumn('users', 'avatar')) {
                 Schema::table('users', function (Blueprint $table) {
                     $table->string('avatar')->after('user_idcode')->default('');
                 });
             }
-            
-            if (Schema::hasColumn('user_profiles','avatar')) {
+
+            if (Schema::hasColumn('user_profiles', 'avatar')) {
                 Schema::table('user_profiles', function (Blueprint $table) {
                     $table->dropColumn('avatar');
-                });     
-            }   
+                });
+            }
         }
 
         $this->tablePerTenant('users', function (Blueprint $table) {
             $table->string('avatar')->after('user_idcode')->default('');
-        },'avatar');
+        }, 'avatar');
 
         $this->tablePerTenant('user_profiles', function (Blueprint $table) {
             $table->dropColumn('avatar');
-        },'avatar',true);
+        }, 'avatar', true);
     }
 
     /**
@@ -46,11 +46,15 @@ class MoveAvatarColumnToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('avatar');
-        });
-        Schema::table('user_profiles', function (Blueprint $table) {
-            $table->string('avatar')->after('user_id')->default('');
-        });
+        if (Schema::hasColumn('users', 'avatar')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('avatar');
+            });
+        }
+        if (!Schema::hasColumn('user_profiles', 'avatar')) {
+            Schema::table('user_profiles', function (Blueprint $table) {
+                $table->string('avatar')->after('user_id')->default('');
+            });
+        }
     }
 }
